requirements.txt
tensorflow==2.19.1
keras
opencv-python
deepface
numpy
requests
pytube    # if YouTube
flask     # if you want a web app
spotipy
python-dotenv
youtube-search-python
